requirements.txt
# requirements.txt

blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-cors==4.0.0
Flask-graphql==2.0.1
flask-sqlalchemy==3.1.1
graphene==2.1.9
graphene-sqlalchemy==2.3.0
gunicorn==21.2.0
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
werkzeug==3.0.1
zipp==3.18.1